doc: document TLSClientFactory
It seems TLSClientFactory works regardless of whether we're
using a proxy or not.

Therefore, using the TLSClientFactory field directly is the
simplest way to use this library with uTLS.

Hence, update the examples. The basic example now uses
the TLSClientFactory. I've copied the original example to
be a more advanced usage example.

While there, also update the proxy example to use the
per-transport TLSClientFactory rather than the global one.

Also, make sure the README file is in sync with the
recently implemented TLSClientFactory changes.
