+++ example/example-utls-with-dial/main.go
@@ -1,15 +1,11 @@
 package main
 
 import (
-	"crypto/tls"
 	"flag"
 	"fmt"
 	"io"
 	"log"
-	"net"
 	"net/http"
-
-	oohttp "github.com/ooni/oohttp"
 )
 
 // newClient creates a new http.Client using the given transport.
@@ -17,18 +13,16 @@ func newClient(txp http.RoundTripper) *http.Client {
 	return &http.Client{Transport: txp}
 }
 
+// defaultClient is the default http.Client
+var defaultClient = newClient(defaultTransport)
+
 func main() {
 	// The default URL we use should return us the JA3 fingerprint
 	// we're using to communicate with the server. We expect such a
 	// fingerprint to change when we use the `-utls` flag.
 	url := flag.String("url", "https://ja3er.com/json", "the URL to get")
-	utls := flag.Bool("utls", false, "try using uTLS")
 	flag.Parse()
-	var ffun func(conn net.Conn, config *tls.Config) oohttp.TLSConn
-	if *utls {
-		ffun = utlsFactory
-	}
-	resp, err := newClient(newTransport(ffun)).Get(*url)
+	resp, err := defaultClient.Get(*url)
 	if err != nil {
 		log.Fatal(err)
 	}
