+++ go.mod
@@ -2,7 +2,6 @@ module github.com/ooni/oohttp
 
 go 1.17
 
-require (
-	golang.org/x/net v0.0.0-20211205041911-012df41ee64c
-	golang.org/x/text v0.3.7 // indirect
-)
+require golang.org/x/net v0.0.0-20220114011407-0dd24b26b47d
+
+require golang.org/x/text v0.3.7 // indirect
